refactor(ci): remove `_command` suffix
