+++ tools/ci/src/commands/clippy.rs
