+++ tools/ci/src/commands/compile_fail.rs
