+++ tools/ci/src/commands/doc.rs
