+++ tools/ci/src/commands/doc_check.rs
