+++ tools/ci/src/commands/example_check.rs
