+++ tools/ci/src/commands/format.rs
