+++ tools/ci/src/commands/mod.rs
@@ -1,29 +1,29 @@
-pub(crate) use bench_check_command::*;
-pub(crate) use cfg_check_command::*;
-pub(crate) use clippy_command::*;
-pub(crate) use compile_check_command::*;
-pub(crate) use compile_command::*;
-pub(crate) use compile_fail_command::*;
-pub(crate) use doc_check_command::*;
-pub(crate) use doc_command::*;
-pub(crate) use doc_test_command::*;
-pub(crate) use example_check_command::*;
-pub(crate) use format_command::*;
-pub(crate) use lints_command::*;
-pub(crate) use test_check_command::*;
-pub(crate) use test_command::*;
+pub(crate) use bench_check::*;
+pub(crate) use cfg_check::*;
+pub(crate) use clippy::*;
+pub(crate) use compile_check::*;
+pub(crate) use compile::*;
+pub(crate) use compile_fail::*;
+pub(crate) use doc_check::*;
+pub(crate) use doc::*;
+pub(crate) use doc_test::*;
+pub(crate) use example_check::*;
+pub(crate) use format::*;
+pub(crate) use lints::*;
+pub(crate) use test_check::*;
+pub(crate) use test::*;
 
-mod bench_check_command;
-mod cfg_check_command;
-mod clippy_command;
-mod compile_check_command;
-mod compile_command;
-mod compile_fail_command;
-mod doc_check_command;
-mod doc_command;
-mod doc_test_command;
-mod example_check_command;
-mod format_command;
-mod lints_command;
-mod test_check_command;
-mod test_command;
+mod bench_check;
+mod cfg_check;
+mod clippy;
+mod compile_check;
+mod compile;
+mod compile_fail;
+mod doc_check;
+mod doc;
+mod doc_test;
+mod example_check;
+mod format;
+mod lints;
+mod test_check;
+mod test;
