+++ tools/ci/src/commands/test.rs
